[Wallet]: Fix Confirmation Panel Size in Side Panel
