+++ components/brave_wallet_ui/components/extension/confirm-transaction-panel/confirm-transaction-panel.stories.tsx
@@ -4,7 +4,7 @@
 // you can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as React from 'react'
-import { PanelWrapper } from '../../../panel/style'
+import { PanelWrapper } from '../../../panel/panel_wrapper/panel_wrapper'
 import { LongWrapper } from '../../../stories/style'
 import WalletPanelStory from '../../../stories/wrappers/wallet-panel-story-wrapper'
