+++ components/brave_wallet_ui/components/extension/connections/connections.stories.tsx
@@ -13,7 +13,7 @@ import { Connections } from './connections'
 
 // Styled Components
 import { LongWrapper } from '../../../stories/style'
-import { PanelWrapper } from '../../../panel/style'
+import { PanelWrapper } from '../../../panel/panel_wrapper/panel_wrapper'
 
 export const _Connections = {
   render: () => {
