+++ components/brave_wallet_ui/components/extension/edit-gas/edit-gas.stories.tsx
@@ -21,7 +21,7 @@ import {
   mockTransactionInfo, //
 } from '../../../stories/mock-data/mock-transaction-info'
 import { mockSepolia } from '../../../stories/mock-data/mock-networks'
-import { PanelWrapper } from '../../../panel/style'
+import { PanelWrapper } from '../../../panel/panel_wrapper/panel_wrapper'
 
 export const _EditGas = {
   render: () => {
