+++ components/brave_wallet_ui/components/extension/post-confirmation/cancel_transaction/cancel_transaction.stories.tsx
@@ -17,7 +17,7 @@ import { CancelTransaction } from './cancel_transaction'
 
 // Styled Components
 import { LongWrapper } from '../../../../stories/style'
-import { PanelWrapper } from '../../../../panel/style'
+import { PanelWrapper } from '../../../../panel/panel_wrapper/panel_wrapper'
 
 export const _CancelTransaction = {
   render: () => {
