+++ components/brave_wallet_ui/components/extension/post-confirmation/complete/complete.stories.tsx
@@ -25,7 +25,7 @@ import { TransactionComplete } from './complete'
 
 // Styled Components
 import { LongWrapper } from '../../../../stories/style'
-import { PanelWrapper } from '../../../../panel/style'
+import { PanelWrapper } from '../../../../panel/panel_wrapper/panel_wrapper'
 
 export const _TransactionComplete = {
   render: (args: StorybookTransactionArgs) => {
